chore: move available jobs related code
