+++ packages/apps/human-app/frontend/src/modules/worker/jobs/available-jobs/components/desktop/index.ts
@@ -1,0 +1,1 @@
+export * from './available-jobs-table-desktop';
