+++ packages/apps/human-app/frontend/src/modules/worker/jobs/available-jobs/components/mobile/available-jobs-drawer-mobile.tsx
@@ -7,10 +7,10 @@ import CloseIcon from '@mui/icons-material/Close';
 import type { Dispatch, SetStateAction } from 'react';
 import { HumanLogoIcon } from '@/shared/components/ui/icons';
 import { useHandleMainNavIconClick } from '@/shared/hooks/use-handle-main-nav-icon-click';
-import { AvailableJobsNetworkFilter } from '@/modules/worker/jobs/available-jobs/components/available-jobs-network-filter';
-import { AvailableJobsJobTypeFilter } from '@/modules/worker/jobs/available-jobs/components/available-jobs-job-type-filter';
-import { AvailableJobsRewardAmountSortMobile } from '@/modules/worker/jobs/available-jobs/components/mobile/available-jobs-reward-amount-sort-mobile';
 import { useColorMode } from '@/shared/contexts/color-mode';
+import { AvailableJobsNetworkFilter } from '../available-jobs-network-filter';
+import { AvailableJobsJobTypeFilter } from '../available-jobs-job-type-filter';
+import { AvailableJobsRewardAmountSortMobile } from './available-jobs-reward-amount-sort-mobile';
 
 interface DrawerMobileProps {
   setIsMobileFilterDrawerOpen: Dispatch<SetStateAction<boolean>>;
